remove dialect namespace
